Move hfb utility to separate module to avoid circular imports
